+++ packages/flutter_deck/lib/src/widgets/internal/marker/marker.dart
@@ -1,2 +1,3 @@
 export 'flutter_deck_marker.dart';
+export 'flutter_deck_marker_controller.dart';
 export 'flutter_deck_marker_notifier.dart';